server/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile requirements.in
asgiref==3.8.1
    # via django
certifi==2024.2.2
    # via requests
cffi==1.16.0
charset-normalizer==3.3.2
    # via requests
decorator==5.1.1
    # via retry
django==5.0.4
    # via djangorestframework
djangorestframework==3.15.1
filelock==3.13.4
    # via
    #   huggingface-hub
    #   torch
    #   transformers
fsspec==2024.3.1
    # via
    #   huggingface-hub
    #   torch
huggingface-hub==0.22.2
    # via
    #   sentence-transformers
    #   tokenizers
    #   transformers
idna==3.7
    # via requests
jinja2==3.1.3
    # via torch
joblib==1.4.0
    # via scikit-learn
markupsafe==2.1.5
    # via jinja2
mpmath==1.3.0
    # via sympy
networkx==3.3
    # via torch
numpy==1.26.4
    # via
    #   pgvector
    #   scikit-learn
    #   scipy
    #   sentence-transformers
    #   transformers
packaging==24.0
    # via
    #   huggingface-hub
    #   transformers
pgvector==0.2.5
pillow==10.3.0
    # via sentence-transformers
psycopg2-binary==2.9.9
py==1.11.0
    # via retry
pycparser==2.22
    # via cffi
pyyaml==6.0.1
    # via
    #   huggingface-hub
    #   transformers
regex==2023.12.25
    # via transformers
requests==2.31.0
    # via
    #   huggingface-hub
    #   transformers
retry==0.9.2
safetensors==0.4.3
    # via transformers
scikit-learn==1.4.2
    # via sentence-transformers
scipy==1.13.0
    # via
    #   scikit-learn
    #   sentence-transformers
sentence-transformers==2.6.1
sqlparse==0.5.0
    # via django
sympy==1.12
    # via torch
threadpoolctl==3.4.0
    # via scikit-learn
tokenizers==0.15.2
    # via transformers
torch==2.2.2
    # via sentence-transformers
tqdm==4.66.2
    # via
    #   huggingface-hub
    #   sentence-transformers
    #   transformers
transformers==4.39.3
    # via sentence-transformers
typing-extensions==4.11.0
    # via
    #   huggingface-hub
    #   torch
urllib3==2.2.1
    # via requests
